+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-* Fixes.
+* Update model scene.